add: api.cpp for api endpoint
