+++ html/index.html
@@ -33,6 +33,9 @@
 <br> <br> <br>
 <a href="/login">login endpoint</a>
 <br> <br>
+<br> <br> <br>
+<a href="/api/now">/api/now</a>
+<br> <br>
 
 </body>
 </html>
