+++ srcs/HttpResponse/POST/post_target.cpp
@@ -9,90 +9,6 @@
 #include "StringHandler.hpp"
 #include "MediaType.hpp"
 
-UrlEncodedFormData parse_urlencoded_form_data(const std::vector<unsigned char> &request_body) {
-    UrlEncodedFormData parameters;
-
-    std::vector<std::string> name_value_pairs;
-    std::vector<unsigned char>::const_iterator head, tail;
-    head = request_body.begin();
-    while (head != request_body.end()) {
-        tail = head;
-        while (tail != request_body.end() && *tail != '&') {
-            ++tail;
-        }
-        std::string name_value(head, tail);
-        name_value_pairs.push_back(name_value);
-
-        head = tail;
-        if (head == request_body.end()) {
-            break;
-        }
-        ++head;
-    }
-
-    std::vector<std::string>::const_iterator itr;
-    for (itr = name_value_pairs.begin(); itr != name_value_pairs.end(); ++itr) {
-        const std::string &name_value = *itr;
-        std::size_t delimiter_pos = name_value.find('=');
-        if (delimiter_pos == std::string::npos) {
-            continue;
-        }
-        std::string key = name_value.substr(0, delimiter_pos);
-        std::string value = name_value.substr(delimiter_pos + 1);
-
-        key = StringHandler::decode(key);
-        value = StringHandler::decode(value);
-        parameters[key].push_back(value);
-        std::cout << "key: " << key << ", value: " << value << std::endl;
-    }
-    return parameters;
-}
-
-
-StatusCode HttpResponse::get_urlencoded_form_content() {
-    std::string head = "<!doctype html>\n"
-                       "<html lang=\"ja\">\n"
-                       "<head>\n"
-                       "    <meta charset=\"UTF-8\">\n"
-                       "    <title>POST params</title>\n"
-                       "</head>\n"
-                       "<body>\n";
-
-    std::string tail = "</body>\n"
-                       "</html>";
-
-    UrlEncodedFormData parameters = parse_urlencoded_form_data(this->body_buf_);
-    std::string parameters_html;
-
-    UrlEncodedFormData::const_iterator itr;
-    for (itr = parameters.begin(); itr != parameters.end(); ++itr) {
-        std::ostringstream oss;
-        oss << itr->first << " : ";
-
-        std::vector<std::string> params = itr->second;
-        std::vector<std::string>::const_iterator param;
-        for (param = params.begin(); param != params.end(); ++param) {
-            oss << *param;
-
-            if (param + 1 != params.end()) {
-                oss << ", ";
-            }
-        }
-        std::string escaped_html = HttpMessageParser::escape_html(oss.str());
-        parameters_html.append(escaped_html);
-        parameters_html.append("<br><br>");
-    }
-
-    std::vector<unsigned char> body;
-    body.insert(body.end(), head.begin(), head.end());
-    body.insert(body.end(), parameters_html.begin(), parameters_html.end());
-    body.insert(body.end(), tail.begin(), tail.end());
-    this->body_buf_ = body;
-
-    return StatusOk;
-}
-
-
 std::vector<unsigned char>::iterator get_non_const_itr(std::vector<unsigned char>::iterator begin,
                                                        std::vector<unsigned char>::const_iterator const_itr) {
     typedef std::vector<unsigned char>::const_iterator c_itr;
@@ -338,40 +254,6 @@ StatusCode HttpResponse::upload_multipart_form_data(const std::string &boundary)
 }
 
 
-StatusCode HttpResponse::show_body() {
-    std::string head = "<!doctype html>\n"
-                       "<html lang=\"ja\">\n"
-                       "<head>\n"
-                       "    <meta charset=\"UTF-8\">\n"
-                       "    <title>POST params</title>\n"
-                       "</head>\n"
-                       "<body>\n";
-
-    std::string tail = "</body>\n"
-                       "</html>";
-
-    std::vector<unsigned char> body;
-    body.insert(body.end(), head.begin(), head.end());
-    body.insert(body.end(), this->body_buf_.begin(), this->body_buf_.end());
-    body.insert(body.end(), tail.begin(), tail.end());
-    this->body_buf_ = body;
-
-    return StatusOk;
-}
-
-
-// media-type = type "/" subtype parameters
-// Content-Type: application/x-www-form-urlencoded
-bool HttpResponse::is_urlencoded_form_data() {
-    Result<MediaType, ProcResult> result = this->request_.get_content_type();
-    if (result.is_err()) {
-        return false;
-    }
-    MediaType media_type = result.get_ok_value();
-    return media_type.type() == "application" && media_type.subtype() == "x-www-form-urlencoded";
-}
-
-
 // Content-Type: multipart/form-data; boundary=----WebKitFormBoundaryOzc5oS6JxLwBcmay
 bool HttpResponse::is_multipart_form_data(std::string *boundary) {
     Result<MediaType, ProcResult> result = this->request_.get_content_type();
@@ -395,16 +277,6 @@ bool HttpResponse::is_multipart_form_data(std::string *boundary) {
         *boundary = itr->second;
     }
     return true;
-}
-
-
-StatusCode HttpResponse::show_data() {
-    if (is_urlencoded_form_data()) {
-        // DEBUG_PRINT(YELLOW, "   show_data -> urlencoded_form");
-        return get_urlencoded_form_content();
-    }
-    // DEBUG_PRINT(YELLOW, "   show_data err: 400");
-    return BadRequest;
 }
